fix: KBD-15 인코딩 문제 및 자료형 문제 해결

+++ src/main/java/com/jeonsaeyukjun/jeonsaeyukjunbe/report/Dto/ReportDto.java
@@ -14,7 +14,7 @@ public class ReportDto {
     private String detailAddress;
     private int safetyScore;
     private String lessorName;
-    private int deposit;
+    private long deposit;
     private boolean isDelete;
     private String createAt;
 }
